% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of single-cell to pseudo-bulk data}
\usage{
aggregateData(x, data, fun, scale = FALSE)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{data}{a character string specifying the assay slot to use as input data.}

\item{fun}{a character string specifying the function to use as summary statistic.}

\item{scale}{logical.}
}
\description{
...
}
