% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExprFreq.R
\name{calcExprFreq}
\alias{calcExprFreq}
\title{calcExprFreq}
\usage{
calcExprFreq(x, assay = "counts", th = 0, melt = FALSE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{a character string specifying which assay to use.}

\item{th}{numeric threshold value above which
a gene should be considered to be expressed.}

\item{melt}{logical. Should results be metled into a single data.frame.}
}
\value{
returns a list of length #clusters containing, for each cluster, 
  a data.frame of dimensions #genes x #samples giving the fraction of cells
  that express each gene in each sample. 
  If \code{melt = TRUE}, a data.frame with columns \code{"cluster_id"}, 
  \code{"sample_id"}, and \code{"freq"} will be returned.
}
\description{
Calculates gene expression frequencies
}
\details{
\code{calcExprFreq} computes for each cluster-sample, the fraction of cells
that express a given gene. Here, a gene is considered to be expressed when
the specified measurement value lies above the specified threshold value.
}
\examples{
data(kang)
freqs <- calcExprFreq(kang)
names(freqs)
head(freqs[[1]])

freqs <- calcExprFreq(kang, melt = TRUE)
head(freqs)

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
