% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExprFreqs.R
\name{calcExprFreqs}
\alias{calcExprFreqs}
\title{calcExprFreqs}
\usage{
calcExprFreqs(x, assay = "counts", th = 0)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{a character string specifying which assay to use.}

\item{th}{numeric threshold value above which
a gene should be considered to be expressed.}
}
\value{
returns a list of length #clusters containing, for each cluster, 
  a data.frame of dimensions #genes x #samples giving the fraction of cells
  that express each gene in each sample. 
  If \code{melt = TRUE}, a data.frame with columns \code{"cluster_id"}, 
  \code{"sample_id"}, and \code{"freq"} will be returned.
}
\description{
Calculates gene expression frequencies
}
\details{
\code{calcExprFreq} computes, for each e.g. cluster, cluster-sample, 
or cluster-group, the fraction of cells that express a given gene. 
Here, a gene is considered to be expressed when the specified 
measurement value lies above the specified threshold value.
}
\examples{
data(kang)
frqs <- calcExprFreqs(kang, by = "cluster_id")
head(frqs)

frqs <- calcExprFreqs(kang, by = c("cluster_id", "sample_id"))
names(frqs)
head(frqs[[1]])

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
