% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerGenes.R
\name{findMarkerGenes}
\alias{findMarkerGenes}
\title{Identification of marker-genes via pairwise DE analysis}
\usage{
findMarkerGenes(x, cluster_id, mc.cores = 1, assay = "RNA",
  block = NULL, logfc.threshold = 1, test.use = "wilcox",
  min.pct = 0.25)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{cluster_id}{a character string specifying the clustering to use.
Must match a \code{x@meta.data} column.}

\item{mc.cores}{passed to \code{\link[parallel]{mclapply}}.}

\item{block, lfc, test.use, min.pct}{passed to \code{\link[scran]{findMarkers}}.}
}
\description{
Test for DE expression between all pairs of clusters.
  For each cluster, cluster-specific genes can then be identified 
  by filtering for genes that are DE against most other clusters.
}
