% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flattenPB}
\alias{flattenPB}
\title{flattenPB}
\usage{
flattenPB(pb, do.norm = TRUE)
}
\arguments{
\item{pb}{a pseudo-bulk \code{\link[SingleCellExperiment]{SingleCellExperiment}}
as produced by \code{\link[muscat]{aggregateData}}, with different 
celltypes/clusters are assays.}

\item{do.norm}{Logical; whether to compute a `logcpm` assay.}
}
\value{
A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
}
\description{
Flattens a pseudo-bulk \code{\link[SingleCellExperiment]{SingleCellExperiment}}
as produced by \code{\link[muscat]{aggregateData}} so that all cell types are 
represented in a single assay.
}
