% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffHeatmap.R
\name{plotDiffHeatmap}
\alias{plotDiffHeatmap}
\title{Heatmap of mean-marker expression by cluster-sample}
\usage{
plotDiffHeatmap(x, y, c = NULL, g = NULL, k = NULL, top_n = 20,
  sort_by = "p_adj.loc", decreasing = FALSE, fdr = 0.05, lfc = 1,
  assay = "logcounts", normalize = TRUE, colors = viridis(10))
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{y}{a list of differential testing results 
as returned by \code{\link{runDS}}.}

\item{c}{a character string specifying the comparison 
(contrast of coefficient) for which results should be plotted.
Defaults to the first one available.}

\item{g, k}{character strings specifying the gene and cluster, 
respectively, to include. Defaults to NULL (all genes/clusters).}

\item{top_n}{single numeric specifying the number of genes to include.}

\item{sort_by}{a character string specifying a \code{y} column to sort by.}

\item{fdr}{single numeric specifying the threshold on adjusted p-values
below which results should be considered significant.}

\item{lfc}{single numeric specifying the threshold on absolute logFCs
above which results should be included.}

\item{assay}{a character string specifying which assay 
in \code{assays(x)}to obtain expression values from.}

\item{normalize}{logical specifying whether
mean-expression values be z-normazlied.}

\item{colors}{character vector of colors to use for plotting.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
...
}
\examples{
data(sce)

# compute pseudo-bulk counts
pb <- aggregateData(sce)

# specify design & contrast matrix
ei <- metadata(sce)$experiment_info
design <- model.matrix(~ 0 + ei$group_id)
dimnames(design) <- list(ei$sample_id, levels(ei$group_id))
contrast <- limma::makeContrasts("stim-ctrl", levels = design)

# test for cluster-specific DE 
res <- runDS(sce, pb, design, contrast, method = "edgeR")

}
\author{
Helena L Crowell & Mark D Robinson
}
