% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkerGenes.R
\name{plotMarkerGenes}
\alias{plotMarkerGenes}
\title{Heatmap of mean marker-gene expression by cluster}
\usage{
plotMarkerGenes(x, marker_genes, scale = TRUE, cluster_columns = TRUE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{marker_genes}{a names list of marker genes for each cluster.}

\item{scale}{logical. 
Specifies wether scaled expression values should be scaled 
between 0 and 1 using 1\% and 99\% quantiles as boundaries.}

\item{cluster_columns}{logical. 
Should columns be hierarchically clustered using euclidean distances.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
...
}
\author{
Helena L. Crowell \email{helena.crowells@uzh.ch}.
}
