% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffcyt.R
\name{run_diffcyt}
\alias{run_diffcyt}
\title{...}
\usage{
run_diffcyt(x, group = "group", k = NULL, data = c("counts",
  "normcounts"), log = FALSE, method = c("limma", "LMM"),
  fun = c("mean", "median"))
}
\arguments{
\item{x}{a \code{daFrame}.}

\item{group}{a character string specifying the grouping variable.}

\item{data}{character string specifying the data to use.
Should be one of \code{assayNames(x)}.}

\item{log}{logical. Should log2 be taken before computing means/medians.}

\item{method}{a character string.}

\item{fun}{character string specifying the summary statistic to use.}
}
\description{
...
}
\author{
Helena Lucia Crowell
}
