% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Cluster-specific DE analysis using \code{edgeR}}
\usage{
run_edgeR(x, pb, design, contrast = NULL, coef = NULL,
  min_cells = 10, verbose = TRUE)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{pb}{a named list of pseudo-bulk data for each cluster
computed with \code{aggregateData}.}

\item{design}{a design matrix with row and column names
created with \code{\link[stats]{model.matrix}}.}

\item{contrast}{a matrix of contrasts created with \code{\link[edgeR]{makeContrasts}}.}

\item{coef}{passed to \code{\link[edgeR]{glmQLFTest}}.
Ignored if \code{contrast} is not NULL.}

\item{min_cells}{a numeric. 
Specifies the minimum number of cells in a given cluster-sample 
required to consider the sample for differential testing.}

\item{verbose}{logical. Should information on progress be reported?}

\item{method}{a character string (see details).}
}
\description{
\code{run_edgeR} tests for cluster-specific 
  differential expression by aggregating single-cell 
  measurements and using \code{edgeR} for testing.
}
\details{
\code{run_edgeR} tests for cluster-specific 
  differential expression by aggregating single-cell 
  measurements. Depending on the selected \code{method},
  differential testing is performed on pseudo-bulk data
  obtained via...
  \describe{
  \item{\code{raw_counts}}{
    summing, for every gene, raw counts for each cluster-sample.}
  \item{\code{normed_counts}}{
    summing, for every gene, normalized counts for each cluster-sample.}
  \item{\code{scaled_cpm}}{
    summing, for every gene, scaled CPM for each cluster-sample.
    Scaled CPM are obtained by multiplying pseudo-bulk raw counts
    by effective library sizes and dividing by 1M.}}
}
\author{
Helena Lucia Crowell \email{helena@crowells.eu}
}
