% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDS.R
\name{runDS}
\alias{runDS}
\title{Cluster-specific DE analysis using}
\usage{
runDS(x, pb, design, contrast = NULL, coef = NULL,
  method = c("edgeR", "limma"), min_cells = 10, verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{pb}{a named list of pseudo-bulk data for each cluster
computed with \code{aggregateData}.}

\item{design}{a design matrix with row and column names
created with \code{\link[stats]{model.matrix}}.}

\item{contrast}{a matrix of contrasts created with \code{\link[edgeR]{makeContrasts}}.}

\item{coef}{passed to \code{\link[edgeR]{glmQLFTest}}.
Ignored if \code{contrast} is not NULL.}

\item{method}{a character string.}

\item{min_cells}{a numeric. 
Specifies the minimum number of cells in a given cluster-sample 
required to consider the sample for differential testing.}

\item{verbose}{logical. Should information on progress be reported?}
}
\value{
a list containing 
\itemize{
\item a data.frame with differential testing results,
\item a \code{\link[edgeR]{DGEList}} object of length nb.-clusters, and
\item the \code{design} matrix, and \code{contrast} or \code{coef} used.
}
}
\description{
\code{run_edgeR} tests for cluster-specific 
  differential expression by aggregating single-cell 
  measurements and using \code{edgeR} for testing.
}
\details{
\code{run_edgeR} tests for cluster-specific 
  differential expression by aggregating single-cell 
  measurements. Depending on the selected \code{method},
  differential testing is performed on pseudo-bulk data
  obtained via...
  \describe{
  \item{\code{raw_counts}}{
    summing, for every gene, raw counts for each cluster-sample.}
  \item{\code{normed_counts}}{
    summing, for every gene, normalized counts for each cluster-sample.}
  \item{\code{scaled_cpm}}{
    summing, for every gene, scaled CPM for each cluster-sample.
    Scaled CPM are obtained by multiplying pseudo-bulk raw counts
    by effective library sizes and dividing by 1M.}}
}
\examples{
# simulate 5 clusters, 20\% of DE genes
data(kang)
sim <- simData(kang, n_genes = 100, n_cells = 200, 
    p_dd = c(0.8, 0, 0.2, 0, 0, 0), fc = 4)
    
# compute pseudo-bulk counts
pb <- aggregateData(sim)

# specify design & contrast matrix
ei <- metadata(sim)$experiment_info
design <- model.matrix(~ 0 + ei$group)
dimnames(design) <- list(ei$sample_id, levels(ei$group))
contrast <- limma::makeContrasts("B-A", levels = design)

# test for cluster-specific DE 
res <- runDS(sim, pb, design, contrast, method = "edgeR")

names(res)
names(res[[1]])
lapply(res[[1]]$`B-A`, head)

# count nb. of DE genes by cluster
n_de <- sapply(res[[1]]$`B-A`, function(x) sum(x$FDR < 0.05))

# get top_n hits for ea. cluster w/ abs(logFC) > 1
top_n <- 5
lapply(res[[1]]$`B-A`, function(x) {
  x <- x[abs(x$logFC) > 1, ]
  x[order(x$FDR)[seq_len(top_n)], ]
})

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
