% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData}
\usage{
simData(
  x,
  ng = nrow(x),
  nc = 2000,
  ns = 3,
  nk = 3,
  probs = NULL,
  p_dd = diag(6)[1, ],
  paired = FALSE,
  p_ep = 0.5,
  p_dp = 0.3,
  p_dm = 0.5,
  p_type = 0,
  lfc = 2,
  rel_lfc = NULL,
  cells_phylo = NULL,
  params_dist = c(0.2, 3)
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{ng, nc, ns, nk}{# of genes, cells, samples and clusters to simulate.}

\item{probs}{a list of length 3 containing probabilities of a cell belonging
to each cluster, sample, and group, respectively. List elements must be 
NULL (equal probabilities) or numeric values in [0, 1] that sum to 1.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{paired}{logial specifying whether a paired design should 
be simulated (both groups use the same set of reference samples) 
or not (reference samples are drawn at random).}

\item{p_ep, p_dp, p_dm}{numeric specifying the proportion of cells
to be shifted to a different expression state in one group (see details).}

\item{p_type}{numeric. Probaility of EE/EP gene being a type-gene.
If a gene is of class "type" in a given cluster, a unique mean 
will be used for that gene in the respective cluster.}

\item{lfc}{numeric value to use as mean logFC
for DE, DP, DM, and DB type of genes.}

\item{rel_lfc}{numeric vector of relative logFCs for each cluster. 
Should be of length \code{nlevels(x$cluster_id)} with 
\code{levels(x$cluster_id)} as names. 
Defaults to factor of 1 for all clusters.}

\item{cells_phylo}{newick tree text representing cluster relations 
and their relative distance. If a tree is given, the distance between 
the clusters will be translated in the number of shared genes 
(this relation is controlled with \code{params_dist}). The distance 
between two clusters is defined as the sum of the branches lengths 
separating them.}

\item{params_dist}{numeric vector of length 2. Defines the number of shared 
genes as an adaptation of the exponential's PDF:
N = Ngenes x gamma1 * e^(-gamma2 x dist) ,
where gamma1 is the parameter that controls the percentage of shared genes. 
By default, it corresponds to \code{p_type} but it's advised to tune it 
depending on the input prep_sce. gamma2 is the 'penalty' of increasing 
the distance between clusters, applied on the number of shared genes. 
Default to -3.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
  containing multiple clusters & samples across 2 groups.
}
\description{
Simulation of complex scRNA-seq data
}
\details{
\code{simData} simulates multiple clusters and samples 
across 2 experimental conditions from a real scRNA-seq data set.
}
\examples{
data(sce)
library(SingleCellExperiment)

# prep. SCE for simulation
sce <- prepSim(sce)

# simulate data
(sim <- simData(sce, ng = 100, nc = 10,
  p_dd = c(0.9, 0, 0.1, 0, 0, 0)))

# simulation metadata
head(gi <- metadata(sim)$gene_info)

# should be ~10\% DE  
table(gi$category)

# unbalanced sample sizes
sim <- simData(sce, ng = 10, nc = 100,
  probs = list(NULL, c(0.25, 0.75), NULL))
table(sim$sample_id)

# one group only
sim <- simData(sce, ng = 10, nc = 100,
  probs = list(NULL, NULL, c(1, 0)))
levels(sim$group_id)


# Hierarchical cell-type relations: 
# we first define a phylogram representing the relations between 3 clusters
cells_phylo <- "(('cluster1':0.1, 'cluster2':0.1):0.4,'cluster3':0.5);"
# to verify the syntax and the relation, we can plot it
library(phylogram)
dend <- read.dendrogram(text = cells_phylo)
plot(dend)
# More complex relations are also possible
cells_phylo2 <- "(('cluster1':0.4, 'cluster2':0.4):0.4, ('cluster3':0.5,('cluster4':0.2,'cluster5':0.2,'cluster6':0.2):0.4):0.4);"
dend2 <- read.dendrogram(text = cells_phylo2)
plot(dend2)
# simulate clusters based on these distances
sim <- simData(sce_, nk = 3, cells_phylo = cells_phylo)
# the information about the shared 'type' genes are also kept in the metadata
table(metadata(sim)$gene_info$shared_class)

}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Helena L Crowell
}
