% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDD.R
\name{simDD}
\alias{simDD}
\title{Simulation of complex scRNA-seq data}
\usage{
simDD(x, fit = NULL, n_genes = 200, n_cells = c(50, 100),
  p_dd = c(0.4, 0.35, 0.12, 0.08, 0.04, 0.01), seed = 1)
}
\arguments{
\item{x}{a \code{SummarizedExperiment}.}

\item{fit}{a NB fit to \code{x}.}

\item{n_genes}{# of genes to simulate.}

\item{n_cells}{# of cells to simulate. 
Either a single numeric or range to sample from.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{seed}{random seed.}
}
\description{
`simDD` simulates multiple clusters and samples 
  across 2 experimental conditions from a real scRNA-seq data set.
}
\examples{
data(kang_se, kang_fit)
simDD(kang_se, kang_fit,
    n_genes = 10, n_cells = 10,
    p_dd = c(1,0,0,0,0,0), seed = 1)

}
