% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of single-cell to pseudo-bulk data}
\usage{
aggregateData(x, data, fun = c("sum", "mean", "median"), scale = FALSE)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{data}{a character string. 
Specifies the assay slot to use as input data.}

\item{fun}{a character string.
Specifies the function to use as summary statistic.}

\item{scale}{logical.}
}
\value{
a list of sample-wise pseudo-bulk data for each cluster.
}
\description{
...
}
\author{
Helena L. Crowel \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
