% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of single-cell to pseudo-bulk data}
\usage{
aggregateData(x, assay, fun = c("sum", "mean", "median"),
  scale = FALSE)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{a character string. 
Specifies the assay slot to use as input data.}

\item{fun}{a character string.
Specifies the function to use as summary statistic.}

\item{scale}{logical.}
}
\value{
a list of sample-wise pseudo-bulk data for each cluster.
}
\description{
...
}
\examples{
data(kang)

pb <- aggregateData(kang, assay = "counts", fun = "sum")
names(pb)
head(pb[[1]])

counts <- assay(kang)
assays(kang)$cpm <- edgeR::cpm(counts)
pb <- aggregateData(kang, assay = "cpm", fun = "sum", scale = TRUE)
head(pb[[1]])

}
\author{
Helena L. Crowel \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
