% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of single-cell to pseudo-bulk data}
\usage{
aggregateData(x, assay = NULL, by = c("cluster_id", "sample_id"),
  fun = c("sum", "mean", "median"), scale = FALSE)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{character string specifying the assay slot to use as 
input data. Defaults to the 1st available (\code{assayNames(x)[1]}).}

\item{by}{character vector specifying which 
\code{colData(x)} columns to summarize by (at most 2!).}

\item{fun}{a character string.
Specifies the function to use as summary statistic.}

\item{scale}{logical. Should pseudo-bulks be scaled
with the effective library size & multiplied by 1M?}
}
\value{
a \code{[SingleCellExperiment]{SingleCellExperiment}}.
}
\description{
...
}
\examples{
library(SummarizedExperiment)
data(kang)

# pseudo-bulk counts by cluster-sample
pb <- aggregateData(kang)
assayNames(pb)  # one sheet per cluster
head(assay(pb)) # n_genes x n_samples

# scaled CPM
assays(kang)$cpm <- edgeR::cpm(assay(kang))
pb <- aggregateData(kang, assay = "cpm", scale = TRUE)
head(assay(pb)) 

# aggregate by cluster
pb <- aggregateData(kang, by = "cluster_id")
length(assays(pb)) # single assay
head(assay(pb))    # n_genes x n_clusters

}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Helena L Crowell & Mark D Robinson
}
