% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of single-cell to pseudo-bulk data}
\usage{
aggregateData(x, method = c("rawCounts", "normedCounts", "scaledCPM"))
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{method}{a character string
specifying the method to use for aggregation.}

\item{cluster_id}{a character string
specifying the column in \code{colData(x)}
that contains cluster assignments.}

\item{sample_id}{a character string
specifying the column in \code{colData(x)}
that contains unique sample identifies.}
}
\description{
...
}
