% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbHW.R
\name{bbhw}
\alias{bbhw}
\title{bbhw: Bulk-based hypothesis weighing}
\usage{
bbhw(
  pbDEA,
  bulkDEA,
  pb = NULL,
  method = c("PAShw", "PAShw.local", "combIHW", "ihw.local", "ihw.global"),
  nbins = NULL,
  alpha = 0.1,
  ...,
  verbose = TRUE,
  BPPARAM = SerialParam(progress = verbose)
)
}
\arguments{
\item{pbDEA}{A data.frame of pseudo-bulk DEA results, as for instance 
produced by \code{\link{pbDS}} or \code{\link{mmDS}} (specifically this 
should be a data.frame for one contrast, e.g. an element of `res$table` of 
the output). This should contain the columns "gene", "cluster_id", "p_val"
and, optionally "logFC".}

\item{bulkDEA}{A data.frame of bulk DEA results, with gene names as row.names
and including the columns "p_val", "gene", "cluster_id" and, ideally,
"logFC". Alternatively, a named vector of significance values. Not that 
these samples should be independent form the single-cell samples on which
`pbDEA` is based.}

\item{pb}{A pseudo-bulk SummarizedExperiment object as produced by 
\code{\link{aggregateData}}.}

\item{method}{The method(s) for which to compiled the adjusted p-values.
By default all methods are computed, but the recommended method is "PAShw"
or, if `pb` is unavailable, `ihw.global`.}

\item{nbins}{The number of significance bins to use for the covariate (i.e. 
prior). For `combIHW`, the effective number of bins will be doubled. We 
recommend leaving this to the default values (which varies across methods).}

\item{alpha}{The nominal level for FDR control for \code{\link[IHW]{ihw}}.}

\item{...}{Passed to \code{\link[IHW]{ihw}}.}

\item{verbose}{Logical; whether to print helpful information.}

\item{BPPARAM}{An optional BiocParallel BPPARAM object for multithreading.}
}
\value{
The `pbDEA` object including extra columns.
}
\description{
This is a method based on Independent Hypothesis Weighing (Ignatiadis et al.,
2016) to increase the power of low-sample-size, per-celltype differential 
state analysis, using a larger dataset of bulk RNAseq.
}
\details{
This function contains the following methods:
* ihw.local and ihw.global simply bin the genes by their significance on the
   bulk RNA-seq, and perform standard independent hypothesis weighing (
   \code{\link[IHW]{ihw}} ) either locally (i.e. for each cell type 
   separately) or globally.
* Proportion-adjusted significance hypothesis weighting (PAShw and 
  PAShw.local) also rely on ihw, but first adjust the covariate according to
  whether the direction of the difference matches (assuming a logFC is 
  available), and then according to how much the cell type contributes to the
  bulk for each specific gene.
* combIHW is the standard global IHW, but the significance bins are further 
  split into low- and high- contributions of the celltype to the bulk, and 
  the bins are treated as nominal.
}
\references{
Ignatiadis, N., Klaus, B., Zaugg, J. et al. Data-driven hypothesis weighting 
increases detection power in genome-scale multiple testing. Nat Methods 13, 
577–580 (2016). https://doi.org/10.1038/nmeth.3885
}
\author{
Pierre-Luc Germain
}
