% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExprFreqs.R
\name{calcExprFreqs}
\alias{calcExprFreqs}
\title{calcExprFreqs}
\usage{
calcExprFreqs(x, assay = "counts", th = 0)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{a character string specifying which assay to use.}

\item{th}{numeric threshold value above which
a gene should be considered to be expressed.}
}
\value{
a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
  containing, for each cluster, an assay of dimensions #genes x #samples 
  giving the fraction of cells that express each gene in each sample. 
  If \code{colData(x)} contains a \code{"group_id"} column, the fraction 
  of expressing cells in each each group will be included as well.
}
\description{
Calculates gene expression frequencies
}
\details{
\code{calcExprFreq} computes, for each e.g. cluster, cluster-sample, 
or cluster-group, the fraction of cells that express a given gene. 
Here, a gene is considered to be expressed when the specified 
measurement value lies above the specified threshold value.
}
\examples{
library(SummarizedExperiment)

colnames(colData(kang)) # contains sample_id only
frq <- calcExprFreqs(kang)
assayNames(frq)  # one assay per cluster
head(assay(frq)) # expr. freqs. by sample

sim <- simData(kang, n_genes = 5, n_cells = 1e3) 
colnames(colData(sim)) # contains group_id!
frq <- calcExprFreqs(sim)
assay(frq) # expr. freqs. by group included

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
