% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmDS.R, R/utils-mmDS.R
\name{mmDS}
\alias{mmDS}
\alias{.mm_dream}
\alias{.mm_vst}
\title{DS analysis using mixed-models (MM)}
\usage{
mmDS(x, coef = NULL, covs = NULL, method = c("dream", "vst"),
  n_cells = 10, n_samples = 2, min_count = 1, min_cells = 20,
  n_threads = 8, verbose = TRUE, ...)

.mm_dream(x, coef, covs, n_threads, verbose, dup_corr = FALSE,
  trended = TRUE, ddf = c("Satterthwaite", "Kenward-Roger"))

.mm_vst(x, coef, covs, n_threads, verbose, blind = TRUE, REML = TRUE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{coef}{character specifying the coefficient to test. 
If NULL (default), will test the last level of \code{"group_id"}.}

\item{covs}{character vector of \code{colData(x)} 
column names to use as covariates.}

\item{method}{a character string. 
Either \code{"dream"} (default, lme4 with voom-weights), 
\code{"vst"} (lme4 with variance-stabilizing transformation), 
or a function accepting the same arguments.}

\item{n_cells}{number of cells per cluster-sample 
required to consider a sample for testing.}

\item{n_samples}{number of samples per group 
required to consider a cluster for testing.}

\item{min_count}{numeric. For a gene to be tested in a given cluster, 
at least \code{min_cells} must have a count >= \code{min_count}.}

\item{min_cells}{number (or fraction, if < 1) of cells with a count >
\code{min_count} required for a gene to be tested in a given cluster.}

\item{n_threads}{number of threads to use.}

\item{verbose}{logical specifying whether messages 
on progress and a progress bar should be displayed.}

\item{...}{additional parameters passed to the method function.}

\item{dup_corr}{logical; whether to use 
\code{\link[limma]{duplicateCorrelation}}.}

\item{trended}{logical; whether to use expression-dependent variance priors
in \code{\link[limma]{eBayes}}.}

\item{ddf}{character string specifying the method for estimating 
the effective degrees of freedom. Either \code{"Satterthwaite"}
(faster) or \code{"Kenward-Roger"} (more accurate) or . 
See \code{\link[variancePartition]{dream}} for details.}

\item{blind}{logical; whether to ignore experimental design for the vst.}

\item{REML}{logical; whether to maximize REML instead of log-likelihood.}
}
\value{
a data.frame
}
\description{
Performs cluster-wise DE analysis by fitting cell-level models.
}
\section{Functions}{
\itemize{
\item \code{.mm_dream}: A SCE wrapper around the voom-lme4-implementation 
\code{\link[variancePartition]{dream}} of mixed models for RNAseq data. 
\code{.mm_dream} expects cells from a single cluster, and
does not perform filtering or handle incorrect parameters well.
Meant to be called by \code{mmDS} with \code{method = "dream"}
to be applied across all clusters.

\item \code{.mm_vst}: A SCE wrapper around \code{\link[DESeq2]{varianceStabilizingTransformation}} 
followed by \code{lme4} mixed models. 
\code{.mm_vst} expects cells from a single cluster, and 
does not perform filtering or handle incorrect parameters well. 
Meant to be called by \code{mmDS} with \code{method = "vst"}
to be applied across all clusters.
}}

\author{
Pierre-Luc Germain and Helena L. Crowell.
}
