% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbMDS.R
\name{pbMDS}
\alias{pbMDS}
\title{Pseudo-bulk level MDS plot}
\usage{
pbMDS(x)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}} 
containing pseudo-bulk data as returned by \code{\link{aggregateData}}.}
}
\value{
a \code{ggplot} object.
}
\description{
...
}
\examples{
sce <- toyData()
sce <- prepSCE(sce, "cluster_id", "sample_id", "group_id")
pb <- aggregateData(sce)
pbMDS(pb)

}
\author{
Helena L. Crowell \email{helena.crowells@uzh.ch} and Mark D. Robinson.
}
