% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffGenes.R
\name{plotDiffGenes}
\alias{plotDiffGenes}
\title{Heatmap of mean-marker expression by cluster-sample}
\usage{
plotDiffGenes(x, y, c = NULL, g = NULL, k = NULL, top_n = 20,
  sort_by = c("FDR", "logFC"), fdr = 0.05, lfc = 1)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{y}{a list of differential testing results 
as returned by \code{\link{runDS}}.}

\item{c}{a character string specifying the comparison 
(contrast of coefficient) for which results should be plotted.
Defaults to the first one available.}

\item{g, k}{character strings specifying the gene and cluster, 
respectively, to include. Defaults to NULL (all genes/clusters).}

\item{top_n}{single numeric specifying the number of genes to include.}

\item{sort_by}{a character string specifying a \code{y} column to sort by.}

\item{fdr}{single numeric specifying the threshold on adjusted p-values
below which results should be considered significant.}

\item{lfc}{single numeric specifying the threshold on absolute logFCs
above which results should be included.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
...
}
\author{
Helena L. Crowell \email{helena@crowells.eu}
}
