% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare SCE for DS analysis}
\usage{
prepData(x, cluster_id, sample_id, group_id)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{cluster_id, sample_id, group_id}{character strings specifying
the \code{colData(x)} columns containing cluster assignments,
unique sample identifiers, and group IDs (e.g., treatment).}
}
\value{
a \code{[SingleCellExperiment]{SingleCellExperiment}}.
}
\description{
...
}
\examples{
# generate random counts
n_genes <- 50
n_cells <- 100
counts <- matrix(
    sample(n_genes * n_cells),
    nrow = n_genes, ncol = n_cells)
    
# generate some cell metadata
sids <- sample(paste0("sample", 1:5), n_cells, TRUE)  # sample IDs
cids <- sample(paste0("cluster", 1:5), n_cells, TRUE) # cluster IDs
condition <- sample(c("ctrl", "stim"), n_cells, TRUE) # conditions
batch <- sample(1:3, n_cells, TRUE)                   # batch nb.

# construct SCE
library(SingleCellExperiment)
sce <- SingleCellExperiment(
    assays = list(counts = counts),
    colData = data.frame(id = sids, cluster = cids, batch, condition))
    
# prep. for workflow
sce <- prepData(sce, 
    cluster_id = "cluster", 
    sample_id = "id", 
    group_id = "condition")
    
head(colData(sce))
metadata(sce)$experiment_info

}
