% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare SCE for DS analysis}
\usage{
prepData(x, cluster_id, sample_id, factors)
}
\arguments{
\item{x}{a \code{[SingleCellExperiment]{SingleCellExperiment}}.}

\item{cluster_id}{a character string
specifying the column in \code{colData(x)}
that contains cluster assignments.}

\item{sample_id}{a character string
specifying the column in \code{colData(x)}
that contains unique sample identifies.}

\item{factors}{a character string
specifying columns  in \code{colData(x)}
that containfactors of interest.}
}
\description{
...
}
