% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSCE.R
\name{prepSCE}
\alias{prepSCE}
\title{Prepare SCE for DS analysis}
\usage{
prepSCE(
  x,
  kid = "cluster_id",
  sid = "sample_id",
  gid = "group_id",
  drop = FALSE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{kid, sid, gid}{character strings specifying
the \code{colData(x)} columns containing cluster assignments,
unique sample identifiers, and group IDs (e.g., treatment).}

\item{drop}{logical. Specifies whether \code{colData(x)} columns
besides those specified as \code{cluster_id,sample_id,group_id}
should be retained (default \code{drop = FALSE}) 
or removed (\code{drop = TRUE}).}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}
\description{
...
}
\examples{
# generate random counts
ng <- 50
nc <- 200
counts <- matrix(sample(ng * nc), nrow = ng, ncol = nc)
    
# generate some cell metadata
gids <- sample(c("groupA", "groupB"), nc, TRUE)   
sids <- sample(paste0("sample", seq_len(3)), nc, TRUE) 
kids <- sample(paste0("cluster", seq_len(5)), nc, TRUE) 
batch <- sample(seq_len(3), nc, TRUE)

# construct SCE
library(SingleCellExperiment)
sce <- SingleCellExperiment(
  assays = list(counts = counts),
  colData = data.frame(group = gids, id = sids, cluster = kids, batch))
    
# prep. for workflow
sce <- prepSCE(sce,
  group_id = "group",
  sample_id = "id", 
  cluster_id = "cluster")
    
head(colData(sce))
metadata(sce)$experiment_info

}
\author{
Helena L Crowell
}
