% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDS.R
\name{resDS}
\alias{resDS}
\title{resDS
Formatting of DS analysis results}
\usage{
resDS(x, y, bind = c("col", "row"), frq = FALSE, cpm = FALSE,
  digits = 3, ...)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{y}{a list of cluster-level differential testing results 
as returned by \code{\link{runDS}}.}

\item{bind}{character string specifying the output format (see details).}

\item{frq}{logical or a pre-computed list of expression frequencies 
as returned by \code{\link{clacExprFreqs}}.}

\item{cpm}{logical specifying whether CPM by cluster-sample 
should be appendeded to the output result table(s).}

\item{digits}{integer value specifying the 
number of significant digits to maintain.}

\item{...}{optional arguments passed to 
\code{\link{clacExprFreqs}} if \code{frq = TRUE}.}
}
\value{
returns a data.frame.
}
\description{
\code{resDS} provides a simple wrapper to format cluster-level
differential testing results into an easily filterable table, and
to optionally append gene expression frequencies by cluster-sample
& -group, as well as cluster-sample-wise CPM.
}
\details{
When \code{bind = "col"}, the list of DS testing results at 
  \code{y$table} will be merge vertically (by column) into a single table
  in tidy format with column \code{contrast/coef} specifying the comparison.
  
  Otherwise, when \code{bind = "row"}, an identifier of the respective
  contrast or coefficient will be appended to the column names,
  and all tables will be merge horizontally (by row).
  
  Expression frequencies pre-computed with \code{\link{calcExprFreqs}} 
  may be provided with \code{frq}. Alternatively, when \code{frq = TRUE}, 
  expression frequencies can be computed directly, and additional arguments 
  may be passed to \code{\link{calcExprFreqs}} (see examples below).
}
\examples{
# simulate 5 clusters, 20\% of DE genes
data(kang)
sim <- simData(kang, n_genes = 100, n_cells = 200, 
  p_dd = c(0.8, 0, 0.2, 0, 0, 0), fc = 4)

# compute pseudo-bulk counts
pb <- aggregateData(sim, assay = "counts", fun = "sum")

# specify design & contrast matrix
ei <- S4Vectors::metadata(sim)$experiment_info
design <- stats::model.matrix(~ 0 + ei$group_id)
dimnames(design) <- list(ei$sample_id, levels(ei$group_id))
contrast <- limma::makeContrasts("B-A", levels = design)

# test for cluster-specific DE 
res <- runDS(sim, pb, design, contrast, method = "edgeR")

head(resDS(sim, res, bind = "row"))
head(resDS(sim, res, bind = "col", digits = Inf))

head(resDS(sim, res, cpm = TRUE)) # append CPM by sample
head(resDS(sim, res, frq = TRUE)) # append expr. freqs. by sample

# pre-computed expr. freqs. & append
frq <- calcExprFreqs(sim, th = 10)
head(resDS(sim, res, frq = frq))

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch} and Mark D. Robinson.
}
