% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAST.R
\name{runMAST}
\alias{runMAST}
\title{Run DS analysis using MAST}
\usage{
runMAST(x, formula, contrast, assay = "logcpm", covs = NULL)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{formula}{model formula.}

\item{contrast}{a contrast matrix 
created with \code{\link[limma]{makeContrasts}}.}

\item{assay}{character string specifying 
which assay to use as input data.}

\item{covs}{character vector of \code{colData(x)} 
column names to use as covariates.}
}
\value{
a list containing \itemize{
\item \code{table}: a list of differential testing results 
for each contrast and cluster, and 
\item the \code{formula} and \code{contrast} used.}
}
\description{
...
}
\examples{

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch}
}
