% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData}
\usage{
simData(x, n_genes = 500, n_cells = 300, probs = NULL,
  p_dd = diag(6)[1, ], fc = 2)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{n_genes}{# of genes to simulate.}

\item{n_cells}{# of cells to simulate. 
Either a single numeric or a range to sample from.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{fc}{numeric value to use as mean logFC
for DE, DP, DM, and DB type of genes.}

\item{ns}{nb. of genes common to 1, 2, ..., all clusters.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
  containing multiple clusters & samples across 2 groups.
}
\description{
Simulation of complex scRNA-seq data
}
\details{
\code{simData} simulates multiple clusters and samples 
across 2 experimental conditions from a real scRNA-seq data set.
}
\examples{
data(kang)
simData(kang,
    n_genes = 10, n_cells = 10,
    p_dd = c(1,0,0,0,0,0))

}
