% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData}
\usage{
simData(
  x,
  nc = 2000,
  ns = 3,
  nk = 3,
  probs = NULL,
  p_dd = diag(6)[1, ],
  paired = FALSE,
  p_ep = 0.5,
  p_dp = 0.3,
  p_dm = 0.5,
  p_type = 0,
  lfc = 2,
  rel_lfc = NULL,
  phylo_tree = NULL,
  phylo_pars = c(ifelse(is.null(phylo_tree), 0, 0.1), 3),
  ng = nrow(x),
  force = FALSE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{nc, ns, nk}{# of cells, samples and clusters to simulate.}

\item{probs}{a list of length 3 containing probabilities of a cell belonging
to each cluster, sample, and group, respectively. List elements must be 
NULL (equal probabilities) or numeric values in [0, 1] that sum to 1.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{paired}{logial specifying whether a paired design should 
be simulated (both groups use the same set of reference samples) 
or not (reference samples are drawn at random).}

\item{p_ep, p_dp, p_dm}{numeric specifying the proportion of cells
to be shifted to a different expression state in one group (see details).}

\item{p_type}{numeric. Probability of EE/EP gene being a type-gene.
If a gene is of class "type" in a given cluster, a unique mean 
will be used for that gene in the respective cluster.}

\item{lfc}{numeric value to use as mean logFC
for DE, DP, DM, and DB type of genes.}

\item{rel_lfc}{numeric vector of relative logFCs for each cluster. 
Should be of length \code{nlevels(x$cluster_id)} with 
\code{levels(x$cluster_id)} as names. 
Defaults to factor of 1 for all clusters.}

\item{phylo_tree}{newick tree text representing cluster relations 
and their relative distance. An explanation of the syntax can be found 
\href{http://evolution.genetics.washington.edu/phylip/newicktree.html}{here}. 
The distance between the nodes, except for the original branch, will be 
translated in the number of shared genes between the clusters belonging to 
these nodes (this relation is controlled with \code{phylo_pars}). The distance 
between two clusters is defined as the sum of the branches lengths 
separating them.}

\item{phylo_pars}{vector of length 2 providing the parameters that control 
the number of type genes. Passed to an exponential's PDF:
\code{N = #genes x gamma1 * e^(-gamma2 x dist)},

\itemize{
\item  \code{gamma1} is the parameter that controls the percentage of shared 
genes between the nodes. By default 0.1 if a tree is given, meaning that 
maximum 10% of the genes can be used as type genes (if gamma2 = 0). 
However it's advised to tune it depending on the input \code{prep_sce}. 
\code{gamma2} is the 'penalty' of increasing the distance between clusters 
(\code{dist}, defined by \code{phylo_tree}), applied on the number of 
shared genes. Default to 3. 
}}

\item{ng}{# of genes to simulate. Importantly, for the library sizes 
computed by \code{\link{prepSim}} (= \code{exp(x$offset)}) to make sense, 
the number of simulated genes should match with the number of genes 
in the reference. To simulate a reduced number of genes, e.g. for 
testing and development purposes, please set \code{force = TRUE}.}

\item{force}{logical specifying whether to force 
simulation despite \code{ng != nrow(x)}.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
  containing multiple clusters & samples across 2 groups.
}
\description{
Simulation of complex scRNA-seq data
}
\details{
\code{simData} simulates multiple clusters and samples 
across 2 experimental conditions from a real scRNA-seq data set.

The simulation of type genes can be performed in 2 ways; 
  (1) by defining \code{p_type} and thus simulating independant clusters, OR 
  (2) by defining both \code{phylo_tree} and \code{phylo_pars}, which will 
  simulate a hierarchical structure between the clusters.
}
\examples{
data(sce)
library(SingleCellExperiment)

# prep. SCE for simulation
ref <- prepSim(sce)

# simulate data
(sim <- simData(ref, nc = 10,
  p_dd = c(0.9, 0, 0.1, 0, 0, 0),
  ng = 100, force = TRUE,
  probs = list(NULL, NULL, c(1, 0))))

# simulation metadata
head(gi <- metadata(sim)$gene_info)

# should be ~10\% DE  
table(gi$category)

# unbalanced sample sizes
sim <- simData(ref, nc = 100, ns = 2,
  probs = list(NULL, c(0.25, 0.75), NULL),
  ng = 10, force = TRUE)
table(sim$sample_id)

# one group only
sim <- simData(ref, nc = 100,
  probs = list(NULL, NULL, c(1, 0)),
  ng = 10, force = TRUE)
levels(sim$group_id)

# HIERARCHICAL CLUSTER STRUCTURE
# define phylogram specifying cluster relations
phylo_tree <- "(('cluster1':0.1,'cluster2':0.1):0.4,'cluster3':0.5);"
# verify syntax & visualize relations
library(phylogram)
plot(read.dendrogram(text = phylo_tree))

# let's use a more complex phylogeny
phylo_tree <- "(('cluster1':0.4,'cluster2':0.4):0.4,('cluster3':
  0.5,('cluster4':0.2,'cluster5':0.2,'cluster6':0.2):0.4):0.4);"
plot(read.dendrogram(text = phylo_tree))

# simulate clusters accordingly
sim <- simData(ref, phylo_tree = phylo_tree, 
  phylo_pars = c(0.1, 3), 
  ng = 500, force = TRUE)
# view information about shared 'type' genes
table(rowData(sim)$class)

}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Helena L Crowell
}
