% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData}
\usage{
simData(x, n_genes, n_cells, p_dd, fc = 2, seed = 1)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{n_genes}{# of genes to simulate.}

\item{n_cells}{# of cells to simulate. 
Either a single numeric or a range to sample from.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{seed}{random seed.}

\item{ns}{nb. of genes common to 1, 2, ..., all clusters.}
}
\description{
Simulation of complex scRNA-seq data
}
\details{
\code{simData} simulates multiple clusters and samples 
across 2 experimental conditions from a real scRNA-seq data set.
}
\examples{
data(kang_se, kang_fit)
simDD(kang_se, kang_fit,
    n_genes = 10, n_cells = 10,
    p_dd = c(1,0,0,0,0,0), seed = 1)

}
