% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData}
\usage{
simData(x, n_genes = 500, n_cells = 300, probs = NULL,
  p_dd = diag(6)[1, ], p_type = 0, lfc = 1, rel_lfc = NULL)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{n_genes}{# of genes to simulate.}

\item{n_cells}{# of cells to simulate. 
Either a single numeric or a range to sample from.}

\item{probs}{a list of length 3 containing probabilities of a cell belonging
to each cluster, sample, and group, respectively. List elements must be 
NULL (equal probabilities) or numeric values in [0, 1] that sum to 1.}

\item{p_dd}{numeric vector of length 6.
Specifies the probability of a gene being
EE, EP, DE, DP, DM, or DB, respectively.}

\item{p_type}{numeric. Probaility of EE/EP gene being a type-gene.
If a gene is of class "type" in a given cluster, a unique mean 
will be used for that gene in the respective cluster.}

\item{lfc}{numeric value to use as mean logFC
for DE, DP, DM, and DB type of genes.}

\item{rel_lfc}{numeric vector of relative logFCs for each cluster. 
Should be of length \code{nlevels(x$cluster_id)} with 
\code{levels(x$cluster_id)} as names. 
Defaults to factor of 1 for all clusters.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
  containing multiple clusters & samples across 2 groups.
}
\description{
Simulation of complex scRNA-seq data
}
\details{
\code{simData} simulates multiple clusters and samples 
across 2 experimental conditions from a real scRNA-seq data set.
}
\examples{
data(sce)
library(SingleCellExperiment)

# prep. SCE for simulation
sce <- prepSim(sce)

# simulate data
(sim <- simData(sce,
  n_genes = 100, n_cells = 10,
  p_dd = c(0.9, 0, 0.1, 0, 0, 0)))

# simulation metadata
head(gi <- metadata(sim)$gene_info)

# should be ~10\% DE  
table(gi$category)

# unbalanced sample sizes
sim <- simData(sce,
  n_genes = 10, n_cells = 100,
  probs = list(NULL, c(0.1, 0.3, 0.6), NULL))
table(sim$sample_id)

# one group only
sim <- simData(sce,
  n_genes = 10, n_cells = 100,
  probs = list(NULL, NULL, c(1, 0)))
levels(sim$group_id)
    
}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Helena L Crowell
}
